+++ core/src/factories/MapperFactory.js
@@ -3,6 +3,7 @@
 import AOROMMapper from '../mappers/AOROMMapper';
 import BNROMMapper from '../mappers/BNROMMapper';
 import CNROMMapper from '../mappers/CNROMMapper';
+import ColorDreamsMapper from '../mappers/ColorDreamsMapper';
 import MMC1Mapper from '../mappers/MMC1Mapper';
 import MMC3Mapper from '../mappers/MMC3Mapper';
 import NINA001Mapper from '../mappers/NINA001Mapper';
@@ -18,6 +19,7 @@ var mappers = {
   'MMC3': MMC3Mapper,
   'AOROM': AOROMMapper,
   'BNROM': BNROMMapper,
+  'ColorDreams': ColorDreamsMapper,
   'NINA-001': NINA001Mapper,
 };